Python/Python/bot/requirements.txt
discord.py==1.7.3
async-generator==1.10
async-timeout==3.0.1